requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
altair>=5.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0